% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird-page.R
\name{birdPage}
\alias{birdPage}
\title{Create a page with a top level navigation bar}
\usage{
birdPage(
  title,
  ...,
  id = NULL,
  selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL,
  footer = NULL,
  inverse = FALSE,
  collapsible = FALSE,
  fluid = TRUE,
  responsive = NULL,
  windowTitle = title,
  color = "blue",
  background_color = "#fafbfc",
  font = NULL,
  theme = NULL
)
}
\arguments{
\item{title}{The title to display in the app}

\item{...}{[tabPanel()] elements to include in the page. The
`navbarMenu` function also accepts strings, which will be used as menu
section headers. If the string is a set of dashes like `"----"` a
horizontal separator will be displayed in the menu.}

\item{id}{If provided, you can use `input$`*`id`* in your
server logic to determine which of the current tabs is active. The value
will correspond to the `value` argument that is passed to
[tabPanel()].}

\item{selected}{The `value` (or, if none was supplied, the `title`)
of the tab that should be selected by default. If `NULL`, the first
tab will be selected.}

\item{position}{Determines whether the navbar should be displayed at the top
of the page with normal scrolling behavior (`"static-top"`), pinned at
the top (`"fixed-top"`), or pinned at the bottom
(`"fixed-bottom"`). Note that using `"fixed-top"` or
`"fixed-bottom"` will cause the navbar to overlay your body content,
unless you add padding, e.g.: \code{tags$style(type="text/css", "body
{padding-top: 70px;}")}}

\item{header}{Tag or list of tags to display as a common header above all
tabPanels.}

\item{footer}{Tag or list of tags to display as a common footer below all
tabPanels}

\item{inverse}{`TRUE` to use a dark background and light text for the
navigation bar}

\item{collapsible}{`TRUE` to automatically collapse the navigation
elements into a menu when the width of the browser is less than 940 pixels
(useful for viewing on smaller touchscreen device)}

\item{fluid}{`TRUE` to use a fluid layout. `FALSE` to use a fixed
layout.}

\item{responsive}{This option is deprecated; it is no longer optional with
Bootstrap 3.}

\item{windowTitle}{The title that should be displayed by the browser window.
Useful if `title` is not a string.}

\item{color}{Optional color for the app. It can be red (default), blue ... or
a valid hex color such as #e0245e.}

\item{background_color}{Optional background color for the app. Any valid hex
color such as #fafbfc.}

\item{font}{Optional Google Font for a quick customization.}

\item{theme}{Optional CSS file within the 'www' directory.}
}
\description{
Create a page that contains a top level navigation bar that can be used to
toggle a set of [tabPanel()] or [tabPanelWithTitle()] elements.
}
\examples{
birdPage(id = "Doge", "Very Shiny")

}
