% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front-page.R
\name{frontPage}
\alias{frontPage}
\title{Create a styled front page}
\usage{
frontPage(title = "My Shiny App", subtitle = NULL, background = NULL, ...)
}
\arguments{
\item{title}{The title to display in the front page}

\item{subtitle}{The subtitle to display in the front page}

\item{background}{The background to display in the front page}

\item{...}{additional arguments}
}
\description{
Create a front page with title, subtitle and background image.
}
\examples{
frontPage("Much content", "Very nice")

}
