% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{frontTitle}
\alias{frontTitle}
\title{Suit Layout Front Title}
\usage{
frontTitle(title, subtitle = NULL)
}
\arguments{
\item{title}{text string for the tab title}

\item{subtitle}{text string for the tab subtitle}
}
\value{
a tab panel for Shiny
}
\description{
Suit Layout Front Title
}
