% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{generateColors}
\alias{generateColors}
\title{Tints of User Defined Color}
\usage{
generateColors(color, pct)
}
\arguments{
\item{color}{complete}

\item{pct}{complete}
}
\value{
tint of specified color in hex format
}
\description{
Tints of User Defined Color
}
\keyword{internal}
