% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{introTitle}
\alias{introTitle}
\title{Suit Layout Introduction Title}
\usage{
introTitle(title)
}
\arguments{
\item{title}{text string for the tab title}
}
\value{
a tab panel for Shiny
}
\description{
Suit Layout Introduction Title
}
