% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{suitBoxLarge}
\alias{suitBoxLarge}
\title{Large Box for Suit Page}
\usage{
suitBoxLarge(title, text, image = NULL, ..., color = "red")
}
\arguments{
\item{title}{complete}

\item{text}{complete}

\item{image}{complete}

\item{...}{additional parameters}

\item{color}{complete}
}
\value{
a tab panel for Shiny
}
\description{
Large Box for Suit Page
}
