% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{suitBoxSmall}
\alias{suitBoxSmall}
\title{Small Box for Suit Page}
\usage{
suitBoxSmall(title, text, ..., color = "red")
}
\arguments{
\item{title}{complete}

\item{text}{complete}

\item{...}{additional parameters}

\item{color}{complete}
}
\value{
a tab panel for Shiny
}
\description{
Small Box for Suit Page
}
