% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{suitHead}
\alias{suitHead}
\title{Suit Layout Head}
\usage{
suitHead(..., color = "red")
}
\arguments{
\item{...}{additional arguments}

\item{color}{text string with red/green/blue or a valid hex color, unvalid colors default to red}
}
\value{
a tab panel for Shiny
}
\description{
Suit Layout Head
}
