% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{suitHeader}
\alias{suitHeader}
\title{Suit Layout Header}
\usage{
suitHeader(logo = NULL, href = NULL, caption = NULL)
}
\arguments{
\item{logo}{URL or file name inside www/ for the navbar logo}

\item{href}{link for the navbar logo}

\item{caption}{alternative text for the navbar logo}
}
\value{
a tab panel for Shiny
}
\description{
Suit Layout Header
}
