% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{suitPage}
\alias{suitPage}
\title{Page With a Suit Layout}
\usage{
suitPage(
  title,
  ...,
  id = "dashboard",
  head = NULL,
  header = NULL,
  footer = NULL,
  windowTitle = title
)
}
\arguments{
\item{title}{text string for the tab title}

\item{...}{additional arguments}

\item{id}{page id}

\item{head}{complete}

\item{header}{complete}

\item{footer}{complete}

\item{windowTitle}{complete}
}
\value{
a tab panel for Shiny
}
\description{
Page With a Suit Layout
}
