% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{tabDesc}
\alias{tabDesc}
\title{Styled Description for Tab}
\usage{
tabDesc(description)
}
\arguments{
\item{description}{text string for the tab description}
}
\value{
an html output for Shiny
}
\description{
Styled Description for Tab
}
\examples{
tabDesc("the details about foo bar")
}
