% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-panel-with-title.R
\name{tabPanelWithTitle}
\alias{tabPanelWithTitle}
\title{Tab Panel With Title}
\usage{
tabPanelWithTitle(title, description = NULL, ...)
}
\arguments{
\item{title}{text string for the tab title}

\item{description}{text string for the tab description}

\item{...}{additional arguments}
}
\value{
a tab panel for Shiny
}
\description{
Tab Panel With Title
}
\examples{
tabPanelWithTitle("foo bar", "the details about foo bar")
}
