% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{tabTitle}
\alias{tabTitle}
\title{Styled Title for Tab}
\usage{
tabTitle(title)
}
\arguments{
\item{title}{text string for the tab title}
}
\value{
an html output for Shiny
}
\description{
Styled Title for Tab
}
\examples{
tabTitle("foo bar")
}
