% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front-page.R
\name{frontPage}
\alias{frontPage}
\title{Bird Layout Front Page}
\usage{
frontPage(title = "My Shiny App", subtitle = NULL, background = NULL, ...)
}
\arguments{
\item{title}{The title to display in the front page}

\item{subtitle}{The subtitle to display in the front page}

\item{background}{The background to display in the front page}

\item{...}{additional arguments}
}
\description{
Bird Layout Front Page
}
