% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{suitHead}
\alias{suitHead}
\title{Suit Layout Head}
\usage{
suitHead(
  ...,
  decoration_color = "red",
  background_color = "white",
  background_image = NULL
)
}
\arguments{
\item{...}{additional arguments}

\item{decoration_color}{text string with red/green/blue or a valid hex color, unvalid colors default to red}

\item{background_color}{explain}

\item{background_image}{an image in www/ or an URL}
}
\value{
a tab panel for Shiny
}
\description{
Suit Layout Head
}
