% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-panel-with-title.R
\name{tabPanelWithTitle}
\alias{tabPanelWithTitle}
\title{Create a Styled Panel With Titles}
\usage{
tabPanelWithTitle(title, description = NULL, ...)
}
\arguments{
\item{title}{text string for the tab title}

\item{description}{text string for the tab description}

\item{...}{additional arguments}
}
\description{
Create a panel that works well with dropdowns and adds title and subtitle.
}
\examples{
tabPanelWithTitle("Wow", "Such data")

}
